tests: address review feedback on PC->BC integration tests
- Move pc_xcm_integration to its own file
- Extract shared helpers into tests/common/mod.rs
- Make execute_from return Result<(), InstructionError> via ensure_complete
- Use submodules instead of comment dividers (origin_rejections,
  safe_call_filter, authorize_semantics, refresh, end_to_end)
- Strip 'Scenario N' prefixes and unnecessary comments
- Drop the sibling-sovereign side-effect test (no mechanism for it)
- Document the XCM Outcome::Complete + inner-dispatch-failure semantics
  on the origin/bad-input tests
