+++ runtimes/bulletin-paseo/tests/tests.rs
@@ -16,15 +16,17 @@
 
 #![cfg(test)]
 
+mod common;
+
 use bulletin_paseo_runtime as runtime;
 use bulletin_paseo_runtime::{
 	paseo_constants::{fee::WeightToFee, locations::PeopleLocation},
 	xcm_config::{GovernanceLocation, LocationToAccountId},
 	AllPalletsWithoutSystem, Balances, Block, HopPromotion, Runtime, RuntimeCall, RuntimeEvent,
-	RuntimeGenesisConfig, RuntimeOrigin, SessionKeys, System, TransactionStorage, TxExtension,
-	UncheckedExtrinsic,
+	RuntimeGenesisConfig, RuntimeOrigin, SessionKeys, System, TransactionStorage,
 };
 use bulletin_transaction_storage_primitives::cids::{calculate_cid, CidConfig, HashingAlgorithm};
+use common::{advance_block, assert_ok_ok, construct_and_apply_extrinsic, ALICE};
 use frame_support::{
 	assert_err, assert_ok, dispatch::GetDispatchInfo, pallet_prelude::Hooks, traits::Get,
 };
@@ -33,7 +35,7 @@ use pallet_bulletin_transaction_storage::{
 	AuthorizationExtent, AuthorizationScope, Call as TxStorageCall, Config as TxStorageConfig,
 	Origin as TxStorageOrigin,
 };
-use parachains_common::{AccountId, AuraId, BlockNumber, Hash as PcHash, Signature as PcSignature};
+use parachains_common::{AccountId, AuraId};
 use parachains_runtimes_test_utils::{ExtBuilder, GovernanceOrigin, RuntimeHelper};
 use sp_core::{crypto::Ss58Codec, Encode, Pair};
 use sp_keyring::Sr25519Keyring;
@@ -41,103 +43,11 @@ use sp_runtime::{
 	traits::{TransactionExtension, TxBaseImplication},
 	transaction_validity,
 	transaction_validity::{InvalidTransaction, TransactionSource, TransactionValidityError},
-	ApplyExtrinsicResult, BuildStorage, Either,
+	BuildStorage, Either,
 };
 use std::collections::HashMap;
 use xcm::latest::prelude::*;
 use xcm_runtime_apis::conversions::LocationToAccountHelper;
-
-const ALICE: [u8; 32] = [1u8; 32];
-
-/// Advance to the next block for testing transaction storage.
-fn advance_block() {
-	let current = frame_system::Pallet::<Runtime>::block_number();
-
-	<TransactionStorage as Hooks<_>>::on_finalize(current);
-	<System as Hooks<_>>::on_finalize(current);
-
-	let next = current + 1;
-	System::set_block_number(next);
-
-	frame_system::BlockWeight::<Runtime>::kill();
-	frame_system::BlockSize::<Runtime>::kill();
-
-	<System as Hooks<_>>::on_initialize(next);
-	<TransactionStorage as Hooks<_>>::on_initialize(next);
-}
-
-fn construct_extrinsic(
-	sender: Option<sp_core::sr25519::Pair>,
-	call: RuntimeCall,
-) -> Result<UncheckedExtrinsic, transaction_validity::TransactionValidityError> {
-	// provide a known block hash for the immortal era check
-	frame_system::BlockHash::<Runtime>::insert(0, PcHash::default());
-	let inner = (
-		frame_system::AuthorizeCall::<Runtime>::new(),
-		frame_system::CheckNonZeroSender::<Runtime>::new(),
-		frame_system::CheckSpecVersion::<Runtime>::new(),
-		frame_system::CheckTxVersion::<Runtime>::new(),
-		frame_system::CheckGenesis::<Runtime>::new(),
-		frame_system::CheckEra::<Runtime>::from(sp_runtime::generic::Era::immortal()),
-		frame_system::CheckNonce::<Runtime>::from(if let Some(s) = sender.as_ref() {
-			let account_id = AccountId::from(s.public());
-			frame_system::Pallet::<Runtime>::account(&account_id).nonce
-		} else {
-			0
-		}),
-		frame_system::CheckWeight::<Runtime>::new(),
-		pallet_skip_feeless_payment::SkipCheckIfFeeless::from(
-			pallet_transaction_payment::ChargeTransactionPayment::<Runtime>::from(0u128),
-		),
-		pallet_bulletin_transaction_storage::extension::ValidateStorageCalls::<
-			Runtime,
-			bulletin_paseo_runtime::storage::StorageCallInspector,
-		>::default(),
-		pallet_bulletin_transaction_storage::extension::AllowanceBasedPriority::<Runtime>::default(
-		),
-		frame_metadata_hash_extension::CheckMetadataHash::<Runtime>::new(false),
-	);
-	let tx_ext: TxExtension =
-		cumulus_pallet_weight_reclaim::StorageWeightReclaim::<Runtime, _>::from(inner);
-
-	if let Some(s) = sender.as_ref() {
-		// Signed call.
-		let account_id = AccountId::from(s.public());
-		let payload = sp_runtime::generic::SignedPayload::new(call.clone(), tx_ext.clone())?;
-		let signature = payload.using_encoded(|e| s.sign(e));
-		Ok(UncheckedExtrinsic::new_signed(
-			call,
-			account_id.into(),
-			PcSignature::Sr25519(signature),
-			tx_ext,
-		))
-	} else {
-		// Unsigned call.
-		Ok(UncheckedExtrinsic::new_transaction(call, tx_ext))
-	}
-}
-
-fn construct_and_apply_extrinsic(
-	account: Option<sp_core::sr25519::Pair>,
-	call: RuntimeCall,
-) -> ApplyExtrinsicResult {
-	let dispatch_info = call.get_dispatch_info();
-	let xt = construct_extrinsic(account, call)?;
-	let xt_len = xt.encode().len();
-	tracing::info!(
-		"Applying extrinsic: class={:?} pays_fee={:?} weight={:?} encoded_len={} bytes",
-		dispatch_info.class,
-		dispatch_info.pays_fee,
-		dispatch_info.total_weight(),
-		xt_len
-	);
-	bulletin_paseo_runtime::Executive::apply_extrinsic(xt)
-}
-
-fn assert_ok_ok(apply_result: ApplyExtrinsicResult) {
-	assert_ok!(apply_result);
-	assert_ok!(apply_result.unwrap());
-}
 
 #[test]
 fn transaction_storage_runtime_sizes() {
@@ -1853,520 +1763,4 @@ fn xcm_transact_authorize_account_works() {
 				},
 			);
 		});
-}
-
-// ============================================================================
-// People Chain → Bulletin Chain `authorize_account` integration tests
-//
-// Drives `XcmExecutor::prepare_and_execute` directly with messages shaped the
-// way People Chain sends them (sibling parachain origin, `OriginKind::Xcm`,
-// unpaid execution + Transact). Exercises the full receive-side pipeline:
-// barrier, origin conversion, `SafeCallFilter`, `Authorizer = EnsureXcm<
-// IsSiblingParachain>`, and the pallet's `authorize_account` /
-// `refresh_account_authorization` semantics.
-//
-// Tracks paritytech/polkadot-bulletin-chain#490.
-// ============================================================================
-
-mod pc_xcm_integration {
-	use super::*;
-	use bulletin_paseo_runtime::xcm_config::XcmConfig;
-	use xcm_executor::traits::ConvertLocation;
-
-	/// People Chain location on Paseo. Matches `paseo_constants::PeopleLocation`.
-	fn pc_location() -> Location {
-		PeopleLocation::get()
-	}
-
-	fn auth_period() -> BlockNumber {
-		<<Runtime as TxStorageConfig>::AuthorizationPeriod as Get<BlockNumber>>::get()
-	}
-
-	fn empty() -> AuthorizationExtent {
-		AuthorizationExtent::default()
-	}
-
-	fn extent(
-		bytes: u64,
-		bytes_allowance: u64,
-		transactions: u32,
-		transactions_allowance: u32,
-	) -> AuthorizationExtent {
-		AuthorizationExtent {
-			bytes,
-			bytes_permanent: 0,
-			bytes_allowance,
-			transactions,
-			transactions_allowance,
-		}
-	}
-
-	fn extent_of(who: &AccountId) -> AuthorizationExtent {
-		TransactionStorage::account_authorization_extent(who.clone())
-	}
-
-	/// Build an XCM message in the shape PC uses: free unpaid execution + Transact.
-	fn xcm_transact(call: RuntimeCall, kind: OriginKind) -> Xcm<RuntimeCall> {
-		Xcm::builder_unsafe()
-			.unpaid_execution(Unlimited, None)
-			.transact(kind, None, call.encode())
-			.build()
-	}
-
-	fn execute_from(origin: Location, message: Xcm<RuntimeCall>) -> xcm::latest::Outcome {
-		let mut id = [0u8; 32];
-		xcm_executor::XcmExecutor::<XcmConfig>::prepare_and_execute(
-			origin,
-			message,
-			&mut id,
-			Weight::MAX,
-			Weight::MAX,
-		)
-	}
-
-	fn pc_authorize(who: AccountId, transactions: u32, bytes: u64) -> xcm::latest::Outcome {
-		let call = RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-			who,
-			transactions,
-			bytes,
-		});
-		execute_from(pc_location(), xcm_transact(call, OriginKind::Xcm))
-	}
-
-	fn pc_refresh(who: AccountId) -> xcm::latest::Outcome {
-		let call = RuntimeCall::TransactionStorage(
-			TxStorageCall::<Runtime>::refresh_account_authorization { who },
-		);
-		execute_from(pc_location(), xcm_transact(call, OriginKind::Xcm))
-	}
-
-	fn new_test_ext() -> sp_io::TestExternalities {
-		let mut ext =
-			sp_io::TestExternalities::new(RuntimeGenesisConfig::default().build_storage().unwrap());
-		ext.execute_with(advance_block);
-		ext
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 5 — origin/barrier rejections.
-	// Goal: prove the `Authorizer` is genuinely scoped to siblings (plus root
-	// and the explicit test accounts), not to any old XCM origin.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn relay_chain_origin_cannot_authorize() {
-		new_test_ext().execute_with(|| {
-			let target: AccountId = Sr25519Keyring::Ferdie.to_account_id();
-			let call =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: target.clone(),
-					transactions: 5,
-					bytes: 1_000,
-				});
-
-			let _ = execute_from(Location::parent(), xcm_transact(call, OriginKind::Xcm));
-
-			assert_eq!(extent_of(&target), empty(), "relay-chain origin must not authorize");
-		});
-	}
-
-	#[test]
-	fn sibling_with_sovereign_origin_kind_cannot_authorize() {
-		new_test_ext().execute_with(|| {
-			let target: AccountId = Sr25519Keyring::Ferdie.to_account_id();
-			let call =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: target.clone(),
-					transactions: 5,
-					bytes: 1_000,
-				});
-
-			let _ = execute_from(pc_location(), xcm_transact(call, OriginKind::SovereignAccount));
-
-			assert_eq!(
-				extent_of(&target),
-				empty(),
-				"sibling with OriginKind::SovereignAccount must not authorize",
-			);
-
-			let sovereign = LocationToAccountId::convert_location(&pc_location())
-				.expect("sibling sovereign account must derive");
-			assert_eq!(
-				extent_of(&sovereign),
-				empty(),
-				"derived sibling sovereign must not gain authorization",
-			);
-		});
-	}
-
-	#[test]
-	fn random_local_origin_cannot_authorize() {
-		new_test_ext().execute_with(|| {
-			let target: AccountId = Sr25519Keyring::Ferdie.to_account_id();
-			let stranger_loc =
-				Location::new(0, [Junction::AccountId32 { network: None, id: [0x42u8; 32] }]);
-			let call =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: target.clone(),
-					transactions: 5,
-					bytes: 1_000,
-				});
-
-			let _ = execute_from(stranger_loc, xcm_transact(call, OriginKind::Xcm));
-
-			// `XcmPassthrough` resolves the origin to `pallet_xcm::Origin::Xcm(stranger)`,
-			// which does not match `IsSiblingParachain` and is not in `TestAccounts`,
-			// so the inner `authorize_account` dispatch is rejected with `BadOrigin`.
-			assert_eq!(extent_of(&target), empty());
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 6 — SafeCallFilter.
-	// Storage-mutating calls must not reach dispatch over XCM, even when the
-	// origin is otherwise valid. The filter inspects through `Utility::batch*`.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn sibling_xcm_store_is_blocked() {
-		new_test_ext().execute_with(|| {
-			let who: AccountId = Sr25519Keyring::Alice.to_account_id();
-			assert_ok!(TransactionStorage::authorize_account(
-				RuntimeOrigin::root(),
-				who.clone(),
-				1,
-				1_000
-			));
-
-			let store_call = RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::store {
-				data: vec![0u8; 100],
-			});
-
-			let outcome = execute_from(pc_location(), xcm_transact(store_call, OriginKind::Xcm));
-
-			assert!(
-				outcome.clone().ensure_complete().is_err(),
-				"SafeCallFilter must block sibling-XCM `store`, got: {outcome:?}",
-			);
-			assert_eq!(extent_of(&who), extent(0, 1_000, 0, 1));
-		});
-	}
-
-	#[test]
-	fn sibling_xcm_batch_with_store_is_entirely_blocked() {
-		new_test_ext().execute_with(|| {
-			let target: AccountId = Sr25519Keyring::Bob.to_account_id();
-			let store_target: AccountId = Sr25519Keyring::Alice.to_account_id();
-
-			let authorize_call =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: target.clone(),
-					transactions: 5,
-					bytes: 1_000,
-				});
-			let store_call = RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::store {
-				data: vec![0u8; 50],
-			});
-			let batch_call = RuntimeCall::Utility(pallet_utility::Call::batch {
-				calls: vec![authorize_call, store_call],
-			});
-
-			let outcome = execute_from(pc_location(), xcm_transact(batch_call, OriginKind::Xcm));
-
-			assert!(
-				outcome.clone().ensure_complete().is_err(),
-				"SafeCallFilter must reject the whole batch when any inner call is filtered, got: {outcome:?}",
-			);
-			assert_eq!(
-				extent_of(&target),
-				empty(),
-				"the inner authorize_account must NOT have executed alongside a filtered store",
-			);
-			assert_eq!(extent_of(&store_target), empty());
-		});
-	}
-
-	#[test]
-	fn sibling_xcm_batch_of_only_authorize_calls_succeeds() {
-		new_test_ext().execute_with(|| {
-			let alice: AccountId = Sr25519Keyring::Alice.to_account_id();
-			let bob: AccountId = Sr25519Keyring::Bob.to_account_id();
-
-			let authorize_alice =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: alice.clone(),
-					transactions: 5,
-					bytes: 1_000,
-				});
-			let authorize_bob =
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::authorize_account {
-					who: bob.clone(),
-					transactions: 10,
-					bytes: 2_000,
-				});
-			let batch_call = RuntimeCall::Utility(pallet_utility::Call::batch {
-				calls: vec![authorize_alice, authorize_bob],
-			});
-
-			let outcome = execute_from(pc_location(), xcm_transact(batch_call, OriginKind::Xcm));
-
-			assert!(
-				outcome.clone().ensure_complete().is_ok(),
-				"sibling XCM batch of authorize_account calls must complete, got: {outcome:?}",
-			);
-			assert_eq!(extent_of(&alice), extent(0, 1_000, 0, 5));
-			assert_eq!(extent_of(&bob), extent(0, 2_000, 0, 10));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 2 — additive within an unexpired window.
-	// Headline invariant: PC may issue multiple voucher claims to the same
-	// account in the same period; each claim adds to the existing allowance
-	// and does NOT push expiry forward. Consumed counters are preserved.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn authorize_is_additive_within_window() {
-		new_test_ext().execute_with(|| {
-			let account = Sr25519Keyring::Alice;
-			let who: AccountId = account.to_account_id();
-
-			assert!(pc_authorize(who.clone(), 5, 1_000).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(0, 1_000, 0, 5));
-
-			advance_block();
-			assert_ok_ok(construct_and_apply_extrinsic(
-				Some(account.pair()),
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::store {
-					data: vec![0u8; 200],
-				}),
-			));
-			assert_eq!(extent_of(&who), extent(200, 1_000, 1, 5));
-
-			assert!(pc_authorize(who.clone(), 3, 500).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(200, 1_500, 1, 8));
-
-			assert!(pc_authorize(who.clone(), 2, 250).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(200, 1_750, 1, 10));
-
-			// Expiry must not have been pushed forward by additive claims.
-			let now = System::block_number();
-			System::set_block_number(now + auth_period());
-			assert_eq!(extent_of(&who), empty(), "additive claims must not have extended expiry",);
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 3 — replace after expiry.
-	// On an expired-but-present entry, a new `authorize_account` resets all
-	// consumed counters and replaces (not adds) the allowances.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn authorize_replaces_after_expiry() {
-		new_test_ext().execute_with(|| {
-			let account = Sr25519Keyring::Alice;
-			let who: AccountId = account.to_account_id();
-
-			assert!(pc_authorize(who.clone(), 5, 1_000).ensure_complete().is_ok());
-
-			advance_block();
-			assert_ok_ok(construct_and_apply_extrinsic(
-				Some(account.pair()),
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::store {
-					data: vec![0u8; 400],
-				}),
-			));
-			assert_eq!(extent_of(&who), extent(400, 1_000, 1, 5));
-
-			let now = System::block_number();
-			System::set_block_number(now + auth_period() + 1);
-			assert_eq!(extent_of(&who), empty(), "extent must read empty once expired");
-
-			assert!(pc_authorize(who.clone(), 1, 100).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(0, 100, 0, 1));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 1 — happy path, single account, fresh state.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn authorize_happy_path_from_sibling() {
-		new_test_ext().execute_with(|| {
-			let who: AccountId = Sr25519Keyring::Alice.to_account_id();
-
-			assert!(pc_authorize(who.clone(), 10, 1_000_000).ensure_complete().is_ok());
-
-			assert_eq!(extent_of(&who), extent(0, 1_000_000, 0, 10));
-			assert!(TransactionStorage::account_has_active_authorization(&who));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 4 — independent account scopes.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn account_scopes_are_independent() {
-		new_test_ext().execute_with(|| {
-			let alice: AccountId = Sr25519Keyring::Alice.to_account_id();
-			let bob: AccountId = Sr25519Keyring::Bob.to_account_id();
-
-			assert!(pc_authorize(alice.clone(), 5, 1_000).ensure_complete().is_ok());
-			assert!(pc_authorize(bob.clone(), 10, 2_000).ensure_complete().is_ok());
-
-			assert_eq!(extent_of(&alice), extent(0, 1_000, 0, 5));
-			assert_eq!(extent_of(&bob), extent(0, 2_000, 0, 10));
-
-			let now = System::block_number();
-			System::set_block_number(now + auth_period() + 1);
-			assert_ok!(TransactionStorage::remove_expired_account_authorization(
-				RuntimeOrigin::none(),
-				alice.clone(),
-			));
-			assert_eq!(extent_of(&alice), empty());
-			// Bob's entry is still in storage — re-authorize lands as a
-			// fresh entry rather than failing.
-			assert!(pc_authorize(bob.clone(), 1, 50).ensure_complete().is_ok());
-			assert_eq!(extent_of(&bob), extent(0, 50, 0, 1));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 7 — bad inputs (`bytes: 0`).
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn authorize_with_zero_bytes_fails() {
-		new_test_ext().execute_with(|| {
-			let who: AccountId = Sr25519Keyring::Alice.to_account_id();
-
-			let _ = pc_authorize(who.clone(), 1, 0);
-
-			assert_eq!(extent_of(&who), empty());
-			assert!(!TransactionStorage::account_has_active_authorization(&who));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 8 — `refresh_account_authorization` extends expiry only.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn refresh_extends_only_expiration() {
-		new_test_ext().execute_with(|| {
-			let who: AccountId = Sr25519Keyring::Alice.to_account_id();
-
-			assert!(pc_authorize(who.clone(), 5, 1_000).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(0, 1_000, 0, 5));
-
-			let half = auth_period() / 2;
-			let now = System::block_number();
-			System::set_block_number(now + half);
-
-			assert!(pc_refresh(who.clone()).ensure_complete().is_ok());
-
-			assert_eq!(extent_of(&who), extent(0, 1_000, 0, 5));
-
-			// A block past the *original* expiry must still be active because
-			// refresh extended the window.
-			System::set_block_number(now + auth_period());
-			assert!(
-				TransactionStorage::account_has_active_authorization(&who),
-				"refresh must have extended expiry past the original window",
-			);
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 9 — `refresh_account_authorization` without a prior authorize.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn refresh_without_prior_authorize_fails() {
-		new_test_ext().execute_with(|| {
-			let who: AccountId = Sr25519Keyring::Alice.to_account_id();
-
-			let _ = pc_refresh(who.clone());
-
-			assert_eq!(extent_of(&who), empty());
-			assert!(!TransactionStorage::account_has_active_authorization(&who));
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Scenario 10 — end-to-end: PC authorizes, user stores and renews.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn authorize_then_user_stores_and_renews() {
-		new_test_ext().execute_with(|| {
-			let account = Sr25519Keyring::Alice;
-			let who: AccountId = account.to_account_id();
-
-			// PC issues a 2-tx, 4_000-byte allowance via XCM.
-			assert!(pc_authorize(who.clone(), 2, 4_000).ensure_complete().is_ok());
-			assert_eq!(extent_of(&who), extent(0, 4_000, 0, 2));
-
-			// Authorized user stores 1_000 bytes (feeless, boost-tier).
-			advance_block();
-			let stored_block = System::block_number();
-			assert_ok_ok(construct_and_apply_extrinsic(
-				Some(account.pair()),
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::store {
-					data: vec![0u8; 1_000],
-				}),
-			));
-			assert_eq!(extent_of(&who), extent(1_000, 4_000, 1, 2));
-
-			// Same user renews against the just-stored block/index. `renew`
-			// charges the per-window permanent quota (`bytes_permanent`).
-			advance_block();
-			assert_ok_ok(construct_and_apply_extrinsic(
-				Some(account.pair()),
-				RuntimeCall::TransactionStorage(TxStorageCall::<Runtime>::renew {
-					block: stored_block,
-					index: 0,
-				}),
-			));
-			assert_eq!(
-				extent_of(&who),
-				AuthorizationExtent {
-					bytes: 1_000,
-					bytes_permanent: 1_000,
-					bytes_allowance: 4_000,
-					transactions: 2,
-					transactions_allowance: 2,
-				},
-			);
-		});
-	}
-
-	// ------------------------------------------------------------------------
-	// Origin-vs-target hygiene: authorizing `alice` must not mutate state for
-	// the sibling sovereign account itself. Guards against a class of
-	// origin-confusion bugs where the dispatched origin and the target
-	// account get conflated.
-	// ------------------------------------------------------------------------
-
-	#[test]
-	fn sibling_sovereign_does_not_gain_authorization() {
-		new_test_ext().execute_with(|| {
-			let alice: AccountId = Sr25519Keyring::Alice.to_account_id();
-			let sovereign = LocationToAccountId::convert_location(&pc_location())
-				.expect("sibling sovereign account must derive");
-
-			assert!(pc_authorize(alice.clone(), 5, 1_000).ensure_complete().is_ok());
-
-			assert_eq!(extent_of(&alice), extent(0, 1_000, 0, 5));
-			assert_eq!(
-				extent_of(&sovereign),
-				empty(),
-				"sibling sovereign must not gain authorization as a side effect",
-			);
-		});
-	}
 }
